requirements.txt
# Core
python-telegram-bot==21.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0

# LLM Providers
openai>=1.0.0
google-generativeai==0.8.0

# Database
lancedb==0.15.0
sentence-transformers==3.0.0

# Utilities
python-dotenv==1.0.0
pydantic>=2.7.4
httpx==0.27.0
apscheduler==3.10.0